refactor($LottoGenerator) : lottonumber 패키지 추가
