+++ src/test/java/domain/WinningLottoTest.java
@@ -1,5 +1,6 @@
 package domain;
 
+import domain.lottonumber.WinningLotto;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
